\name{Klein}
\alias{Klein}
\docType{data}
\title{
Klein (1950) macro data. 
}
\description{
The data is used to reproduce examples of Greene (2012)
}
\usage{data("Klein")}
\format{
  A data frame with 22 observations on the following 10 variables.
  \describe{
    \item{\code{YEAR}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{P}}{a numeric vector}
    \item{\code{WP}}{a numeric vector}
    \item{\code{I}}{a numeric vector}
    \item{\code{K1}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{WG}}{a numeric vector}
    \item{\code{G}}{a numeric vector}
    \item{\code{T}}{a numeric vector}
  }
}

\source{
  On Greene (2012) online resources.
}
\references{
  Klein, L. (1950), \emph{Economic Fluctuations in the United-States
  1921-1941}, New York: John Wiley and Sons.
  
Green, W.H.. (2012). \emph{Econometric Analysis, 7th edition}, Prentice Hall.    
}
\examples{
data(Klein)
}
\keyword{datasets}
