\name{ManufactCost}
\alias{ManufactCost}
\docType{data}
\title{
  Manufacturing Costs data from Bernt and Wood (1975)
}
\description{
The data is used to reproduce examples of Greene (2012)
}
\usage{data("ManufactCost")}
\format{
  A data frame with 25 observations on the following 10 variables.
  \describe{
    \item{\code{Year}}{a numeric vector}
    \item{\code{Cost}}{a numeric vector}
    \item{\code{K}}{a numeric vector}
    \item{\code{L}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
    \item{\code{M}}{a numeric vector}
    \item{\code{Pk}}{a numeric vector}
    \item{\code{Pl}}{a numeric vector}
    \item{\code{Pe}}{a numeric vector}
    \item{\code{Pm}}{a numeric vector}
  }
}

\source{
  On Greene (2012) online resources.
}
\references{
  Berndt, E. and Wood, D. (1975), \emph{Technology, Prices, and the
    Derived Demand for Energy}, Review of Economics and Statistics,
  Vol. 57, 376--384.
  
Green, W.H.. (2012). \emph{Econometric Analysis, 7th edition}, Prentice Hall.    
}
\examples{
data(ManufactCost)
}
\keyword{datasets}
