\name{minAlgo-class}
\docType{class}
\alias{minAlgo-class}
\alias{minAlgo}

\title{Class \code{"minAlgo"}}
\description{
A union class for all minimization algorithms. It is created by
\code{\link{algoObj}}.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{print}{\code{signature(x = "minAlgo")}: ... }
    \item{show}{\code{signature(object = "minAlgo")}: ... }    
  }
}

\examples{
showClass("minAlgo")
}
\keyword{classes}
