\name{rmomentModel-class}
\docType{class}
\alias{rmomentModel-class}

\title{Class \code{"rmomentModel"}}
\description{
  A union class for all restricted moment-based models.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
  \describe{
    \item{gelFit}{\code{signature(model = "rmomentModel")}: ... }
	 }
}

\examples{
showClass("rmomentModel")
}
\keyword{classes}
