\name{rsysModel-class}
\docType{class}
\alias{rsysModel-class}

\title{Class \code{"rsysModel"}}
\description{
A union class for all systems of equations. (see \code{link{systemGmm}})
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "rsysModel" in the signature.
}

\examples{
showClass("rsysModel")
}
\keyword{classes}
